[NUI] Separate the LayoutTransition from LayoutController
Separate class to help understanding for layout process.
